fix style issues.

+++ cpp/tests/binaryop/integration/assert-binops.h
@@ -41,7 +41,7 @@ void ASSERT_BINOP(cudf::test::column_wrapper<TypeOut>& out,
     auto out_data = std::get<0>(out_h);
 
     ASSERT_TRUE(out_data.size() == rhs_data.size());
-    for (int index = 0; index < out_data.size(); ++index) {
+    for (size_t index = 0; index < out_data.size(); ++index) {
         ASSERT_TRUE(out_data[index] == (TypeOut)(ope(lhs_h, rhs_data[index])));
     }
 
@@ -50,7 +50,7 @@ void ASSERT_BINOP(cudf::test::column_wrapper<TypeOut>& out,
 
     uint32_t lhs_valid = (lhs.is_valid() ? UINT32_MAX : 0);
     ASSERT_TRUE(out_valid.size() == rhs_valid.size());
-    for (int index = 0; index < out_valid.size(); ++index) {
+    for (size_t index = 0; index < out_valid.size(); ++index) {
         ASSERT_TRUE(out_valid[index] == (lhs_valid & rhs_valid[index]));
     }
 }
@@ -67,7 +67,7 @@ void ASSERT_BINOP(cudf::test::column_wrapper<TypeOut>& out,
     auto out_data = std::get<0>(out_h);
 
     ASSERT_TRUE(out_data.size() == lhs_data.size());
-    for (int index = 0; index < out_data.size(); ++index) {
+    for (size_t index = 0; index < out_data.size(); ++index) {
         ASSERT_TRUE(out_data[index] == (TypeOut)(ope(lhs_data[index], rhs_h)));
     }
 
@@ -76,7 +76,7 @@ void ASSERT_BINOP(cudf::test::column_wrapper<TypeOut>& out,
 
     uint32_t rhs_valid = (rhs.is_valid() ? UINT32_MAX : 0);
     ASSERT_TRUE(out_valid.size() == lhs_valid.size());
-    for (int index = 0; index < out_valid.size(); ++index) {
+    for (size_t index = 0; index < out_valid.size(); ++index) {
         ASSERT_TRUE(out_valid[index] == (lhs_valid[index] & rhs_valid));
     }
 }
@@ -95,7 +95,7 @@ void ASSERT_BINOP(cudf::test::column_wrapper<TypeOut>& out,
 
     ASSERT_TRUE(out_data.size() == lhs_data.size());
     ASSERT_TRUE(out_data.size() == rhs_data.size());
-    for (int index = 0; index < (int)out_data.size(); ++index) {
+    for (size_t index = 0; index < out_data.size(); ++index) {
         ASSERT_TRUE(out_data[index] == (TypeOut)(ope(lhs_data[index], rhs_data[index])));
     }
 
@@ -105,7 +105,7 @@ void ASSERT_BINOP(cudf::test::column_wrapper<TypeOut>& out,
 
     ASSERT_TRUE(out_valid.size() == lhs_valid.size());
     ASSERT_TRUE(out_valid.size() == rhs_valid.size());
-    for (int index = 0; index < (int)out_valid.size(); ++index) {
+    for (size_t index = 0; index < out_valid.size(); ++index) {
         ASSERT_TRUE(out_valid[index] == lhs_valid[index] | rhs_valid[index]);
     }
 }
@@ -130,7 +130,7 @@ void ASSERT_BINOP(cudf::test::column_wrapper<TypeOut>& out,
     const int ULP = 2.0;
     ASSERT_TRUE(out_data.size() == lhs_data.size());
     ASSERT_TRUE(out_data.size() == rhs_data.size());
-    for (int index = 0; index < out_data.size(); ++index) {
+    for (size_t index = 0; index < out_data.size(); ++index) {
         ASSERT_TRUE(abs(out_data[index] - (TypeOut)(ope(lhs_data[index], rhs_data[index]))) < ULP);
     }
 
@@ -140,7 +140,7 @@ void ASSERT_BINOP(cudf::test::column_wrapper<TypeOut>& out,
 
     ASSERT_TRUE(out_valid.size() == lhs_valid.size());
     ASSERT_TRUE(out_valid.size() == rhs_valid.size());
-    for (int index = 0; index < out_valid.size(); ++index) {
+    for (size_t index = 0; index < out_valid.size(); ++index) {
         ASSERT_TRUE(out_valid[index] == (lhs_valid[index] & rhs_valid[index]));
     }
 }
